+++ bin/turn_gen.dart
@@ -3,6 +3,6 @@ import 'package:turn_gen/src/src.dart';
 
 void main(List<String> arguments) {
   // ignore: avoid_print
-  print(ConstConsole.introMessage(packageVersion));
+  print(ConstConsole.introMessage());
   turn_gen.runFromArguments(arguments);
 }