+++ lib/src/src.dart
@@ -8,4 +8,3 @@ export 'enum_type_run.dart';
 export 'logger.dart';
 export 'union_class/union_class.dart';
 export 'utils.dart';
-export 'version.dart';